+++ src/rust/ide/lib/view/src/graph_editor.rs
@@ -37,7 +37,7 @@ use ensogl::gui::component::Tween;
 use ensogl::gui::cursor;
 use ensogl::prelude::*;
 use ensogl::system::web;
-use frp::io::keyboard;
+
 
 
 // =================
