+++ src/rust/lib/frp/src/io.rs
@@ -1,8 +1,8 @@
 //! Root module for Input / Output FRP bindings
 
-pub mod keyboard2;
 pub mod keyboard;
+pub mod keyboard_old;
 pub mod mouse;
 
-pub use keyboard::Keyboard;
+pub use keyboard_old::Keyboard;
 pub use mouse::Mouse;
